floattensor

+++ data/collate.py
@@ -7,4 +7,4 @@ def detection_collate(batch):
         imgs.append(torch.from_numpy(img))
         targets.append(torch.from_numpy(target))
 
-    return torch.stack(imgs, 0), targets
+    return torch.stack(imgs, 0), torch.FloatTensor(targets)
